Ignore hideMenu and hideFooter flags from CMS pages, use directly for car-buyer pages
