+++ apps/store/src/app/[locale]/StoryblokLayout.tsx
@@ -21,7 +21,6 @@ export const StoryblokLayout = ({
   // TODO: How should we pass story data here ?  For now, let's pretend it's non-story page
   // Removed breadcrumbs from page router LayoutWithMenu, kept other flags in fixed state
   const story: any = null
-  const hideFooter = Math.PI < 1 // always false, but does not trigger eslint static condition warning
   const overlayMenu = false
 
   const handleLocaleChange = useChangeLocale(story)
@@ -34,9 +33,7 @@ export const StoryblokLayout = ({
   const headerBlock = filterByBlockType(globalStory.content.header, HeaderBlock.blockName)
   const footerBlock = filterByBlockType(globalStory.content.footer, FooterBlock.blockName)
 
-  const showHeader = !story?.content.hideMenu
   const showMenuOverlay = story?.content.overlayMenu ?? overlayMenu
-  const showFooter = !hideFooter && !story?.content.hideFooter
   const darkBackground = story?.content.darkBackground
 
   return (
@@ -44,26 +41,24 @@ export const StoryblokLayout = ({
       {reusableBlock.map((referencedBlok) => (
         <StoryblokComponent key={referencedBlok._uid} blok={referencedBlok} />
       ))}
-      {showHeader &&
-        headerBlock.map((nestedBlock) => (
-          <HeaderBlock
+      {headerBlock.map((nestedBlock) => (
+        <HeaderBlock
+          key={nestedBlock._uid}
+          blok={nestedBlock}
+          overlay={showMenuOverlay}
+          static={story && isProductStory(story)}
+        />
+      ))}
+      {children}
+      {footerBlock.map((nestedBlock) => (
+        <Fragment key={nestedBlock._uid}>
+          <FooterBlock
             key={nestedBlock._uid}
             blok={nestedBlock}
-            overlay={showMenuOverlay}
-            static={story && isProductStory(story)}
+            onLocaleChange={handleLocaleChange}
           />
-        ))}
-      {children}
-      {showFooter &&
-        footerBlock.map((nestedBlock) => (
-          <Fragment key={nestedBlock._uid}>
-            <FooterBlock
-              key={nestedBlock._uid}
-              blok={nestedBlock}
-              onLocaleChange={handleLocaleChange}
-            />
-          </Fragment>
-        ))}
+        </Fragment>
+      ))}
     </div>
   )
 }
