+++ apps/store/src/features/carDealership/CarBuyerCmsPage.tsx
@@ -1,0 +1,14 @@
+import { NextPageWithLayout } from 'next'
+import { LayoutWithMenu } from '@/components/LayoutWithMenu/LayoutWithMenu'
+import { StoryblokPage } from '@/components/StoryblokPage'
+import { StoryblokPageProps } from '@/services/storyblok/storyblok'
+
+export const CarBuyerCmsPage: NextPageWithLayout<StoryblokPageProps> = (props) => (
+  <StoryblokPage {...props} />
+)
+
+CarBuyerCmsPage.getLayout = (children) => (
+  <LayoutWithMenu hideFooter={true} hideMenu={true}>
+    {children}
+  </LayoutWithMenu>
+)
